backend/requirements.txt
# Smart Crédit Requirements
# Install with: pip install -r requirements.txt

# Django
Django==5.0.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9

# IA & LLM
groq==0.7.0
langchain==0.1.8
huggingface-hub==0.20.3

# PDF & Email
WeasyPrint==60.0
Pillow==10.1.0

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
celery==5.3.4
redis==5.0.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Development
black==23.12.1
flake8==6.1.0
isort==5.13.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Monitoring
sentry-sdk==1.39.1
